Moving helpers into a separate file
